// src/gs-api/src/models/EmploiTempDto.ts
import { CourDto } from "./CourDto";

export class EmploiTempDto {
    id?: number;
    salle?: string;
    salle2?: string;
    salle3?: string;
    salle4?: string;
    salle5?: string;
    salle6?: string;
    filiere?: string;
    niveau?: string;
    date?:string;
    cour1?:CourDto;
    cour2?:CourDto;
    cour3?:CourDto;
    cour4?:CourDto;
    cour5?:CourDto;
    cour6?:CourDto;
    cour7?:CourDto;
    cour8?:CourDto;
}